gpu: ion: Disallow non-secure allocations from the CP heap
Currently, the CP heap supports non-secure allocations as long as the
no_nonsecure_alloc flag is not set in the device tree or platform data
for the heap. Since allowing non-secure allocations from secure heaps
is actually the exception rather than the rule, remove this flag and
replace it with a flag that can be set to accomplish the inverse
(i.e. override the policy of disallowing non-secure allocations from
secure heaps).

[mpokwsths: Edited flo board file]

Change-Id: I1d03991a895c3cf2dea67f79d6af0d4244fb5a75
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/board-8064.c
	arch/arm/mach-msm/board-8930.c
	arch/arm/mach-msm/board-8960.c