crypto: msm: Fix kernel panic in sg_virt() call
There exists scenarios, where data is copied from client's
scatter gather list (sglist) to a locally allocated buffer.
In these cases, the driver uses sg_virt() API to retrieve
the virtual address of the buffer for each of the sglist
entry.

Using sg_virt() on scatter list entries, that is designated
for buffers in high memory pages, results in returning a 0
address; NULL pointer is returned.  This is due to the fact
that high memory pages are by definition unmapped. Hence
sg_virt() does not work for such pages.  Dereferencing this
NULL pointer to copy data from/to results in kernel panic.

Fix is to use sg_copy_from_buffer()/sg_copy_to_buffer() to
copy data between a local buffer and the buffer pointed to by
the sglist. These APIs temporarily  maps the memory addresed
by each of the sglist entries using kmap() and copies data
from this mapped segments to the requested local buffer.

Change-Id: Ibaaf218f3f86a35c67368c7b2ee57189db207bca
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>